[v0.86][tools] Split oversized pr_cmd.rs PR lifecycle module into focused units (Closes #1258)
